pyproject.toml
[project]
name = "whispy"
version = "0.1.0"
description = "whisper packages"
authors = [
    {name = "LNWatta",email = "helene.finot@le-campus-numerique.fr"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flask (>=3.1.0,<4.0.0)",
    "celery (>=5.4.0,<6.0.0)",
    "openai-whisper (>=20240930,<20240931)",
    "redis (>=5.2.1,<6.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
